@ReceiveEvent now has priority. Changed Prioritable and PriorityCollection to now operate on floats, rather than ints.

+++ secsy-context/src/main/java/com.gempukku.secsy.context/util/Prioritable.java
@@ -1,5 +1,5 @@
 package com.gempukku.secsy.context.util;
 
 public interface Prioritable {
-    int getPriority();
+    float getPriority();
 }
